convopro_chatgpt/pyproject.toml
[project]
name = "convopro-chatgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index==0.14.0",
    "llama-index-core>=0.14.20",
    "llama-index-llms-ollama==0.7.2",
    "ollama==0.5.3",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "pymongo==4.14.1",
    "python-dotenv==1.1.1",
    "streamlit==1.49.1",
    "watchdog>=6.0.0",
]
